Rename scrapy/dupefilter.py to scrapy/dupefilters.py
